requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
contourpy==1.2.1
cycler==0.12.1
decorator==4.4.2
docopt==0.6.2
einops==0.7.0
filelock==3.13.4
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
greenlet==3.0.3
huggingface-hub==0.22.2
HyperPyYAML==1.2.2
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.0
julius==0.2.7
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.1
lightning==2.2.2
lightning-utilities==0.11.2
llvmlite==0.42.0
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
networkx==3.2.1
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
optuna==3.6.1
packaging==24.0
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.0
pooch==1.8.1
primePy==1.3
proglog==0.1.10
protobuf==5.26.1
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.22
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.2
pytorch-metric-learning==2.5.0
pytube==15.0.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.4.2
scipy==1.13.0
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
speechbrain==1.0.0
SQLAlchemy==2.0.29
sympy==1.12
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.4.0
torch==2.2.2
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.2.2
torchmetrics==1.3.2
tqdm==4.66.2
triton==2.2.0
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
yarl==1.9.4
zipp==3.18.1